# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Это пример сервера с использованием Swagger в Gin.",
        "title": "swagger API",
        "contact": {},
        "version": "1.0"
    },
    "host": "157.90.227.125:8000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/hashtag": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collection"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "hashtag_id",
                        "name": "hashtag_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-model_Hashtag"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-schema_Empty"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hashtag"
                ],
                "parameters": [
                    {
                        "description": "CreatePost data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.UpdateHashtag"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-model_Collection"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-schema_Empty"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "publication"
                ],
                "parameters": [
                    {
                        "description": "CreatePost data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.CreateHashtag"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-model_Hashtag"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-schema_Empty"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hashtag"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "hashtag_id",
                        "name": "hashtag_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-schema_Empty"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-schema_Empty"
                        }
                    }
                }
            }
        },
        "/api/v1/hashtag/add": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hashtag"
                ],
                "summary": "добавить пост в коллекцию",
                "parameters": [
                    {
                        "description": "CreatePost data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.AddHashtag"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-model_Hashtag"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-schema_Empty"
                        }
                    }
                }
            }
        },
        "/api/v1/hashtag/delete-post": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hashtag"
                ],
                "summary": "удалить пост из коллекцию",
                "parameters": [
                    {
                        "description": "CreatePost data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.DeleteHashtagPost"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-model_Hashtag"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-schema_Empty"
                        }
                    }
                }
            }
        },
...
        "/api/v1/stories": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stories"
                ],
                "summary": "получить список сторисов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-array_model_Stories"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-schema_Empty"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stories"
                ],
                "summary": "пометить сторис прочитанным",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "story_page_id",
                        "name": "story_page_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-schema_Empty"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-schema_Empty"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "admin.AddHashtag": {
            "type": "object",
            "properties": {
                "hashtag_id": {
                    "type": "integer"
                },
                "post_id": {
                    "type": "integer"
                }
            }
        },
...
        "admin.CreateHashtag": {
            "type": "object",
            "properties": {
                "image_base64": {
                    "type": "string"
                },
                "is_visible": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "name_kz": {
                    "type": "string"
                },
                "name_ru": {
                    "type": "string"
                }
            }
        },
...
        "admin.DeleteHashtagPost": {
            "type": "object",
            "properties": {
                "hashtag_id": {
                    "type": "integer"
                },
                "post_id": {
                    "type": "integer"
                }
            }
        },
...
        "admin.UpdateHashtag": {
            "type": "object",
            "properties": {
                "hashtag_id": {
                    "type": "integer"
                },
                "image_base64": {
                    "type": "string"
                },
                "is_visible": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "name_kz": {
                    "type": "string"
                },
                "name_ru": {
                    "type": "string"
                }
            }
        },
...
        "model.Collection": {
            "type": "object",
            "properties": {
                "collection_id": {
                    "type": "integer"
                },
                "image_path": {
                    "type": "string"
                },
                "image_path_kz": {
                    "type": "string"
                },
                "image_path_ru": {
                    "type": "string"
                },
                "is_recommendation": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "name_kz": {
                    "type": "string"
                },
                "name_ru": {
                    "type": "string"
                },
                "posts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Post"
                    }
                }
            }
        },
...
        "model.Hashtag": {
            "type": "object",
            "properties": {
                "hashtag_id": {
                    "type": "integer"
                },
                "image_path": {
                    "type": "string"
                },
                "is_visible": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "name_kz": {
                    "type": "string"
                },
                "name_ru": {
                    "type": "string"
                },
                "posts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Post"
                    }
                }
            }
        },
        "model.Image": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "file_name": {
                    "type": "string"
                },
                "image_id": {
                    "type": "integer"
                },
                "post_id": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "model.Post": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "collections": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Collection"
                    }
                },
                "company": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "hashtags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Hashtag"
                    }
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Image"
                    }
                },
                "language": {
                    "type": "string"
                },
                "point": {
                    "type": "integer"
                },
                "post_id": {
                    "type": "integer"
                },
                "post_order": {
                    "type": "integer"
                },
                "quiz_time": {
                    "type": "integer"
                },
                "rating_status": {
                    "type": "string"
                },
                "read_time": {
                    "type": "integer"
                },
                "sapphire": {
                    "type": "integer"
                },
                "short_description": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "model.Stories": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "icon_path": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "stories_id": {
                    "type": "integer"
                },
                "story_pages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.StoryPage"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "model.StoryPage": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "image_path": {
                    "type": "string"
                },
                "is_readed": {
                    "type": "boolean"
                },
                "page_order": {
                    "type": "integer"
                },
                "stories_id": {
                    "type": "integer"
                },
                "story_page_id": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "schema.Empty": {
            "type": "object"
        },
...
        "schema.Response-array_model_Stories": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Stories"
                    }
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "schema.Response-model_Collection": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {
                    "$ref": "#/definitions/model.Collection"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "schema.Response-model_Hashtag": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {
                    "$ref": "#/definitions/model.Hashtag"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "schema.Response-schema_Empty": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {
                    "$ref": "#/definitions/schema.Empty"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}